jp2sql-api/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
pydantic==2.9.2
python-dotenv==1.0.1
asyncpg==0.30.0
sqlparse==0.5.1
google-generativeai>=0.7.0
tenacity==9.0.0
